Fix shoreline initial connection issue (#1)
In some cases, when a client service (i.e. Hydrophone) starts before Shoreline the initial connection to Shoreline will fail but the client service will not exit. This results in a service apparently up and running but not able to commuicate with Shoreline due to the lack of a valid token.
Let's retry to connect to shoreline 5 times before dying and exit with an error.